un-register main suite test in CMakeLists.txt
