+++ AutomatedTesting/Gem/PythonTests/atom_renderer/CMakeLists.txt
@@ -15,23 +15,23 @@
 ################################################################################
 
 if(PAL_TRAIT_BUILD_HOST_TOOLS AND PAL_TRAIT_BUILD_TESTS_SUPPORTED AND AutomatedTesting IN_LIST LY_PROJECTS)
-    ly_add_pytest(
-        NAME AutomatedTesting::AtomRenderer_HydraTests_Main
-        TEST_SUITE main
-        PATH ${CMAKE_CURRENT_LIST_DIR}/test_Atom_MainSuite.py
-        TEST_SERIAL
-        TIMEOUT 300
-        RUNTIME_DEPENDENCIES
-            AssetProcessor
-            AutomatedTesting.Assets
-            Editor
-    )
+#    ly_add_pytest(
+#        NAME AutomatedTesting::AtomRenderer_HydraTests_Main
+#        TEST_SUITE main
+#        PATH ${CMAKE_CURRENT_LIST_DIR}/test_Atom_MainSuite.py
+#        TEST_SERIAL
+#        TIMEOUT 500
+#        RUNTIME_DEPENDENCIES
+#            AssetProcessor
+#            AutomatedTesting.Assets
+#            Editor
+#    )
     ly_add_pytest(
         NAME AutomatedTesting::AtomRenderer_HydraTests_Sandbox
         TEST_SUITE sandbox
         PATH ${CMAKE_CURRENT_LIST_DIR}/test_Atom_SandboxSuite.py
         TEST_SERIAL
-        TIMEOUT 300
+        TIMEOUT 500
         RUNTIME_DEPENDENCIES
             AssetProcessor
             AutomatedTesting.Assets
