+++ AutomatedTesting/Gem/PythonTests/atom_renderer/test_Atom_MainSuite.py
@@ -11,7 +11,16 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 Main suite tests for the Atom renderer.
 """
 
+import logging
+import os
+
 import pytest
+
+import editor_python_test_tools.hydra_test_utils as hydra
+
+logger = logging.getLogger(__name__)
+EDITOR_TIMEOUT = 300
+TEST_DIRECTORY = os.path.join(os.path.dirname(__file__), "atom_hydra_scripts")
 
 
 @pytest.mark.parametrize("project", ["AutomatedTesting"])
@@ -19,6 +28,157 @@ import pytest
 @pytest.mark.parametrize("level", ["auto_test"])
 class TestAtomEditorComponentsMain(object):
 
-    # It requires at least one test
-    def test_Dummy(self, request, editor, level, workspace, project, launcher_platform):
-        pass
+    @pytest.mark.test_case_id(
+        "C32078130",  # Display Mapper
+        "C32078129",  # Light
+        "C32078131",  # Radius Weight Modifier
+        "C32078127",  # PostFX Layer
+        "C32078125",  # Physical Sky
+        "C32078115",  # Global Skylight (IBL)
+        "C32078121",  # Exposure Control
+        "C32078120",  # Directional Light
+        "C32078119",  # DepthOfField
+        "C32078118")  # Decal (Atom)
+    def test_AtomEditorComponents_AddedToEntity(self, request, editor, level, workspace, project, launcher_platform):
+        cfg_args = [level]
+
+        expected_lines = [
+            # Decal (Atom) Component
+            "Decal (Atom) Entity successfully created",
+            "Decal (Atom)_test: Component added to the entity: True",
+            "Decal (Atom)_test: Component removed after UNDO: True",
+            "Decal (Atom)_test: Component added after REDO: True",
+            "Decal (Atom)_test: Entered game mode: True",
+            "Decal (Atom)_test: Exit game mode: True",
+            "Decal (Atom) Controller|Configuration|Material: SUCCESS",
+            "Decal (Atom)_test: Entity is hidden: True",
+            "Decal (Atom)_test: Entity is shown: True",
+            "Decal (Atom)_test: Entity deleted: True",
+            "Decal (Atom)_test: UNDO entity deletion works: True",
+            "Decal (Atom)_test: REDO entity deletion works: True",
+            # DepthOfField Component
+            "DepthOfField Entity successfully created",
+            "DepthOfField_test: Component added to the entity: True",
+            "DepthOfField_test: Component removed after UNDO: True",
+            "DepthOfField_test: Component added after REDO: True",
+            "DepthOfField_test: Entered game mode: True",
+            "DepthOfField_test: Exit game mode: True",
+            "DepthOfField_test: Entity disabled initially: True",
+            "DepthOfField_test: Entity enabled after adding required components: True",
+            "DepthOfField Controller|Configuration|Camera Entity: SUCCESS",
+            "DepthOfField_test: Entity is hidden: True",
+            "DepthOfField_test: Entity is shown: True",
+            "DepthOfField_test: Entity deleted: True",
+            "DepthOfField_test: UNDO entity deletion works: True",
+            "DepthOfField_test: REDO entity deletion works: True",
+            # Exposure Control Component
+            "Exposure Control Entity successfully created",
+            "Exposure Control_test: Component added to the entity: True",
+            "Exposure Control_test: Component removed after UNDO: True",
+            "Exposure Control_test: Component added after REDO: True",
+            "Exposure Control_test: Entered game mode: True",
+            "Exposure Control_test: Exit game mode: True",
+            "Exposure Control_test: Entity disabled initially: True",
+            "Exposure Control_test: Entity enabled after adding required components: True",
+            "Exposure Control_test: Entity is hidden: True",
+            "Exposure Control_test: Entity is shown: True",
+            "Exposure Control_test: Entity deleted: True",
+            "Exposure Control_test: UNDO entity deletion works: True",
+            "Exposure Control_test: REDO entity deletion works: True",
+            # Global Skylight (IBL) Component
+            "Global Skylight (IBL) Entity successfully created",
+            "Global Skylight (IBL)_test: Component added to the entity: True",
+            "Global Skylight (IBL)_test: Component removed after UNDO: True",
+            "Global Skylight (IBL)_test: Component added after REDO: True",
+            "Global Skylight (IBL)_test: Entered game mode: True",
+            "Global Skylight (IBL)_test: Exit game mode: True",
+            "Global Skylight (IBL) Controller|Configuration|Diffuse Image: SUCCESS",
+            "Global Skylight (IBL) Controller|Configuration|Specular Image: SUCCESS",
+            "Global Skylight (IBL)_test: Entity is hidden: True",
+            "Global Skylight (IBL)_test: Entity is shown: True",
+            "Global Skylight (IBL)_test: Entity deleted: True",
+            "Global Skylight (IBL)_test: UNDO entity deletion works: True",
+            "Global Skylight (IBL)_test: REDO entity deletion works: True",
+            # Physical Sky Component
+            "Physical Sky Entity successfully created",
+            "Physical Sky component was added to entity",
+            "Entity has a Physical Sky component",
+            "Physical Sky_test: Component added to the entity: True",
+            "Physical Sky_test: Component removed after UNDO: True",
+            "Physical Sky_test: Component added after REDO: True",
+            "Physical Sky_test: Entered game mode: True",
+            "Physical Sky_test: Exit game mode: True",
+            "Physical Sky_test: Entity is hidden: True",
+            "Physical Sky_test: Entity is shown: True",
+            "Physical Sky_test: Entity deleted: True",
+            "Physical Sky_test: UNDO entity deletion works: True",
+            "Physical Sky_test: REDO entity deletion works: True",
+            # PostFX Layer Component
+            "PostFX Layer Entity successfully created",
+            "PostFX Layer_test: Component added to the entity: True",
+            "PostFX Layer_test: Component removed after UNDO: True",
+            "PostFX Layer_test: Component added after REDO: True",
+            "PostFX Layer_test: Entered game mode: True",
+            "PostFX Layer_test: Exit game mode: True",
+            "PostFX Layer_test: Entity is hidden: True",
+            "PostFX Layer_test: Entity is shown: True",
+            "PostFX Layer_test: Entity deleted: True",
+            "PostFX Layer_test: UNDO entity deletion works: True",
+            "PostFX Layer_test: REDO entity deletion works: True",
+            # Radius Weight Modifier Component
+            "Radius Weight Modifier Entity successfully created",
+            "Radius Weight Modifier_test: Component added to the entity: True",
+            "Radius Weight Modifier_test: Component removed after UNDO: True",
+            "Radius Weight Modifier_test: Component added after REDO: True",
+            "Radius Weight Modifier_test: Entered game mode: True",
+            "Radius Weight Modifier_test: Exit game mode: True",
+            "Radius Weight Modifier_test: Entity is hidden: True",
+            "Radius Weight Modifier_test: Entity is shown: True",
+            "Radius Weight Modifier_test: Entity deleted: True",
+            "Radius Weight Modifier_test: UNDO entity deletion works: True",
+            "Radius Weight Modifier_test: REDO entity deletion works: True",
+            # Light Component
+            "Light Entity successfully created",
+            "Light_test: Component added to the entity: True",
+            "Light_test: Component removed after UNDO: True",
+            "Light_test: Component added after REDO: True",
+            "Light_test: Entered game mode: True",
+            "Light_test: Exit game mode: True",
+            "Light_test: Entity is hidden: True",
+            "Light_test: Entity is shown: True",
+            "Light_test: Entity deleted: True",
+            "Light_test: UNDO entity deletion works: True",
+            "Light_test: REDO entity deletion works: True",
+            # Display Mapper Component
+            "Display Mapper Entity successfully created",
+            "Display Mapper_test: Component added to the entity: True",
+            "Display Mapper_test: Component removed after UNDO: True",
+            "Display Mapper_test: Component added after REDO: True",
+            "Display Mapper_test: Entered game mode: True",
+            "Display Mapper_test: Exit game mode: True",
+            "Display Mapper_test: Entity is hidden: True",
+            "Display Mapper_test: Entity is shown: True",
+            "Display Mapper_test: Entity deleted: True",
+            "Display Mapper_test: UNDO entity deletion works: True",
+            "Display Mapper_test: REDO entity deletion works: True",
+        ]
+
+        unexpected_lines = [
+            "Trace::Assert",
+            "Trace::Error",
+            "Traceback (most recent call last):",
+        ]
+
+        hydra.launch_and_validate_results(
+            request,
+            TEST_DIRECTORY,
+            editor,
+            "hydra_AtomEditorComponents_AddedToEntity.py",
+            timeout=EDITOR_TIMEOUT,
+            expected_lines=expected_lines,
+            unexpected_lines=unexpected_lines,
+            halt_on_unexpected=True,
+            null_renderer=True,
+            cfg_args=cfg_args,
+        )
+
